refactor: Controller에 try~catch 문 반복 코드 Validator로 리팩토링 완료
